added kenton_convenience_functions for testing convenience
