+++ bycycle/kenton_convenience_features/__init__.py
@@ -1,0 +1,3 @@
+"""Burst detection."""
+
+from .simulate_cycles import *
